# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyTaskResponse:
    properties:
      board_id:
        type: string
      board_name:
        type: string
      column_name:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Turtask
  description: |-
    Multi-board Kanban + task management with realtime sync.
    Auth uses an HttpOnly `auth_token` cookie issued by /api/auth/login or /api/auth/oauth.
  title: Turtask API
  version: "1.0"
paths:
...
  /api/my-tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyTaskResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: My tasks (cross-board)
      tags:
      - my-tasks
  /api/my-tasks/{cardID}/complete:
    post:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: not the assignee or card missing
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Complete one of my tasks
      tags:
      - my-tasks
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"